refactor(chat): move ThreadConnection into store/
